app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiokafka==0.12.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
click==7.1.2
colorclass==2.2.2
colorlog==6.9.0
croniter==2.0.7
Faker==30.8.2
faust==1.10.4
faust-streaming==0.11.3
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==1.4.7
MarkupSafe==3.0.2
mode==4.3.2
mode-streaming==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
opentracing==1.3.0
packaging==24.2
propcache==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
robinhood-aiokafka==1.1.6
semi==0.0.4
six==1.16.0
sortedcontainers==2.4.0
terminaltables==3.1.10
toolz==1.0.0
typing_extensions==4.12.2
venusian==3.1.0
Werkzeug==3.1.3
yarl==1.17.1
blinker==1.9.0
click==8.1.7
Faker==30.8.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
toolz==1.0.0
typing_extensions==4.12.2
Werkzeug==3.1.3
